Rename dataclass_builder module -> structured_input; drop legacy names
Completes the Editor/edit rename — the module is no longer dataclass-specific:

- spytial/dataclass_builder.py -> spytial/structured_input.py
- HTML path dataclass_builder() -> edit_html(); internal
  _generate_dataclass_builder_html -> _generate_editor_html
- Clean break: remove the DataClassBuilder / dataclass_builder back-compat
  aliases entirely (no deprecation shim)
- Update __init__ exports, three test files, docs (page renamed to
  structured-input.md; mkdocs nav; api.md mkdocstrings target) and both demos
- demos/04-dataclass-builder.ipynb -> demos/04-structured-input.ipynb

Canonical surface: Editor / edit() (live widget), edit_html() (standalone HTML).
Full suite: 160 passed, 4 skipped.
